AudioExplorer: Refactored code for Import Audio and Cue Edit methods
This patch adds threads to handle audio conversion process and audio
wave file generation. This changes makes the dialog that handle
audio import and cue edit much more responsive and the code is
now much more cleaner. However, there is some more work to be done
to clean-up the code that handles this processes.